+++ lib/groupdate.rb
@@ -20,10 +20,9 @@ module Groupdate
   PERIODS = [:second, :minute, :hour, :day, :week, :month, :quarter, :year, :day_of_week, :hour_of_day, :minute_of_hour, :day_of_month, :day_of_year, :month_of_year]
   METHODS = PERIODS.map { |v| :"group_by_#{v}" } + [:group_by_period]
 
-  mattr_accessor :week_start, :day_start, :time_zone, :dates
+  mattr_accessor :week_start, :day_start, :time_zone
   self.week_start = :sunday
   self.day_start = 0
-  self.dates = true
 
   # api for gems like ActiveMedian
   def self.process_result(relation, result, **options)