+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Changed SQL to return dates instead of times for day, week, month, quarter, and year
 - Added `n` option for Redshift
+- Removed `dates` option
 - Dropped support for Ruby < 2.6 and Rails < 5.2
 
 ## 5.2.4 (2021-12-15)